splitting things up before starting
