+++ lib/adapters/db.js
@@ -1,4 +1,5 @@
 const pgp = require('simple-pg')()
+
 const { database } = require('../config')
 
 module.exports = pgp(database)
