+++ lib/helpers/checkEnum.js
@@ -1,4 +1,4 @@
-module.exports =  function checkEnum (string = '', list = [], defaultValue = null) {
+function checkEnum (string = '', list = [], defaultValue = null) {
   if (typeof string !== 'string') {
     return defaultValue
   }
@@ -11,3 +11,5 @@ module.exports =  function checkEnum (string = '', list = [], defaultValue = nul
 
   return defaultValue
 }
+
+module.exports = checkEnum
