+++ lib/index.js
@@ -1,5 +1,6 @@
 const restify = require('restify')
 const cors = require('cors')
+
 const routes = require('./routes')
 const { port } = require('./config')
