+++ lib/models/Job.js
@@ -1,5 +1,7 @@
 const striptags = require('striptags')
+
 const checkEnum = require('../helpers/checkEnum')
+
 module.exports = class Job {
   static language (string = '') {
     return checkEnum(string, ['en', 'sv', 'ar', 'fa', 'so', 'ti'])
