+++ lib/services/jobs/index.js
@@ -1,0 +1,1 @@
+module.exports = Object.assign({}, require('./get'), require('./insert'))
