+++ lib/services/platforms.js
@@ -1,7 +1,4 @@
 const db = require('../adapters/db')
-const camelcaseKeys = require('camelcase-keys')
-const snakecaseKeys = require('snakecase-keys')
-
 
 async function getPlatforms ({ page = 1, pageLimit = 10 }) {
   const offset = (page - 1) * pageLimit
