+++ test/integration/helpers/elasticClient.js
@@ -2,6 +2,4 @@ const { Client } = require('elasticsearch')
 
 const { elastic: { host } } = require('../config')
 
-module.exports = new Client({
-  host
-})
+module.exports = new Client({ host })
