+++ test/integration/setup.fixture.js
@@ -2,11 +2,7 @@ const { spawn } = require('child_process')
 const { existsSync } = require('fs')
 const chalk = require('chalk')
 
-const log = chalk.keyword('grey')
-const err = chalk.keyword('white')
-
 const client = require('./helpers/psql')('postgres')
-
 const {
   database: {
     user,
@@ -22,6 +18,7 @@ const {
   }
 } = require('./config')
 
+const [log, err] = [chalk.keyword('grey'), chalk.keyword('white')]
 const dbURL = `postgres://${user}:${password}@${host}:${port}`
 
 function migrate ({ args = ['up'], ignoreIfNotExist = false }) {
